plugin dial_mobile: rename service to services
Why:

* comply with wazo-notebook/file-structure.md
